Separating concerns
Tank Service and tank movement service separated.

shader log files still bypassing the gitignore
