+++ .gitignore
@@ -36,3 +36,5 @@ sysinfo.txt
 *.unitypackage
 *.log
 *.log
+Logs/shadercompiler-UnityShaderCompiler.exe0.log
+*.log
