+++ Assets/Prefabs/TankPrefab.prefab
@@ -176,8 +176,9 @@ GameObject:
   - component: {fileID: 7881210056624536841}
   - component: {fileID: 7881210056624536840}
   - component: {fileID: 2366483372870643255}
+  - component: {fileID: 5315554526107036471}
   m_Layer: 0
-  m_Name: Tank
+  m_Name: TankPrefab
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -239,6 +240,19 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5185eb07723c86749b97a96f9752821d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &5315554526107036471
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7881210056624503317}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 46540f3607a3e5549a30e4db358687db, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  input: {fileID: 11400000, guid: b96280f9aa7ea8b44ae072cb75c1b577, type: 2}
 --- !u!1 &7881210056624503319
 GameObject:
   m_ObjectHideFlags: 0
