+++ Assets/Scenes/Game.unity
@@ -1061,10 +1061,9 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   tankView: {fileID: 2366483372870643255, guid: de7989b69a0929644991214c180a0ace,
     type: 3}
-  input: {fileID: 11400000, guid: b96280f9aa7ea8b44ae072cb75c1b577, type: 2}
   _speed: 5
-  _rotateSpeed: 120
-  _jumpForce: 1
+  _rotateSpeed: 360
+  _jumpForce: 0.5
 --- !u!4 &2052177811
 Transform:
   m_ObjectHideFlags: 0
