+++ Assets/Scripts/TankService/TankController.cs
@@ -15,23 +15,13 @@ public class TankController
 
     public void Move(Vector2 _moveDirection)
     {
-        if (_moveDirection == Vector2.zero)
-        {
-            return;
-        }
-        else
-        {
             Vector3 moveDirection = new Vector3(_moveDirection.x, 0, _moveDirection.y);
 
             tankView.MoveFunction(moveDirection, tankModel.Speed, tankModel.RotateSpeed);
-        }
     }
 
-    public void Jump(bool _isJumping)
+    public void Jump()
     {
-        if(_isJumping)
-        {
-            tankView.JumpFunction(tankModel.JumpForce);
-        }
+        tankView.JumpFunction(tankModel.JumpForce);
     }
 };
