+++ Assets/Scripts/TankService/TankMovementService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46540f3607a3e5549a30e4db358687db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
